[이인태] 리포지토리 작성완료, 필요사항 : 멤버, 부서

+++ groovy/src/main/java/com/groovy/ware/calendar/dto/CalendarDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class CalendarDTO {
     
-    private int schCode;
+    private Long schCode;
     private String schTitle;
     private String schContext;
     private String schDiv;
